requirements.txt
selenium
openai
python-dotenv
requests
